src/main/kotlin/com/zenika/handson/spring/repositories/http/Weather7TimerMapper.kt
package com.zenika.handson.spring.repositories.http

import com.zenika.handson.spring.entities.Weather
import java.time.LocalDate
import java.time.format.DateTimeFormatter

class Weather7TimerMapper {

    companion object {

        fun toDomain(dailyWeather7Timer: DailyWeather7Timer): List<Weather> {
            return dailyWeather7Timer.dataseries.map {
                Weather(
                    day = LocalDate.parse(it.date, DateTimeFormatter.ofPattern("yyyyMMdd")),
                    weather = it.weather,
                    temperatureMax = it.temp2m.max,
                    temperatureMin = it.temp2m.min
                )
            }
        }

    }
}
